rename bad funcs instead of removing their samples
Removing the samples for functions that have known inlining issues skews
the PGO profile and might cause previously cold functions to be promoted
to hot functions.

It's probably not a big issue, but this is an easy improvement to make,
so why not.

Kudos to Michael Pratt for pointing this out pointing this out in
https://github.com/DataDog/datadog-pgo/pull/3#discussion_r1579652098